Tidy Teams logic around and improve service functions names
