+++ test/plausible/teams/invitations/invite_to_site_test.exs
@@ -8,21 +8,21 @@ defmodule Plausible.Teams.Invitations.InviteToSiteTest do
 
   @subject_prefix if ee?(), do: "[Plausible Analytics] ", else: "[Plausible CE] "
 
-  describe "create_invitation/4" do
+  describe "invite/4" do
     test "creates an invitation" do
       inviter = new_user()
       invitee = new_user()
       site = new_site(owner: inviter)
 
       assert {:ok, %Plausible.Teams.GuestInvitation{}} =
-               InviteToSite.create_invitation(site, inviter, invitee.email, :viewer)
+               InviteToSite.invite(site, inviter, invitee.email, :viewer)
     end
 
     test "returns validation errors" do
       inviter = new_user()
       site = new_site(owner: inviter)
 
-      assert {:error, changeset} = InviteToSite.create_invitation(site, inviter, "", :viewer)
+      assert {:error, changeset} = InviteToSite.invite(site, inviter, "", :viewer)
       assert {"can't be blank", _} = changeset.errors[:email]
     end
 
@@ -33,10 +33,10 @@ defmodule Plausible.Teams.Invitations.InviteToSiteTest do
       add_guest(site, user: invitee, role: :viewer)
 
       assert {:error, :already_a_member} =
-               InviteToSite.create_invitation(site, inviter, invitee.email, :viewer)
+               InviteToSite.invite(site, inviter, invitee.email, :viewer)
 
       assert {:error, :already_a_member} =
-               InviteToSite.create_invitation(site, inviter, inviter.email, :viewer)
+               InviteToSite.invite(site, inviter, inviter.email, :viewer)
     end
 
     test "sends invitation email for existing users" do
@@ -44,7 +44,7 @@ defmodule Plausible.Teams.Invitations.InviteToSiteTest do
       site = new_site(owner: inviter)
 
       assert {:ok, %Plausible.Teams.GuestInvitation{}} =
-               InviteToSite.create_invitation(site, inviter, invitee.email, :viewer)
+               InviteToSite.invite(site, inviter, invitee.email, :viewer)
 
       assert_email_delivered_with(
         to: [nil: invitee.email],
@@ -57,7 +57,7 @@ defmodule Plausible.Teams.Invitations.InviteToSiteTest do
       site = new_site(owner: inviter)
 
       assert {:ok, %Plausible.Teams.GuestInvitation{invitation_id: invitation_id}} =
-               InviteToSite.create_invitation(site, inviter, "vini@plausible.test", :viewer)
+               InviteToSite.invite(site, inviter, "vini@plausible.test", :viewer)
 
       assert_email_delivered_with(
         to: [nil: "vini@plausible.test"],
@@ -75,7 +75,7 @@ defmodule Plausible.Teams.Invitations.InviteToSiteTest do
       for _ <- 1..4, do: add_guest(site, role: :viewer)
 
       assert {:error, {:over_limit, 3}} =
-               InviteToSite.create_invitation(site, inviter, invitee.email, :viewer)
+               InviteToSite.invite(site, inviter, invitee.email, :viewer)
     end
 
     @tag :ee_only
@@ -84,7 +84,7 @@ defmodule Plausible.Teams.Invitations.InviteToSiteTest do
       site = new_site(owner: owner)
 
       invite = fn site, email ->
-        InviteToSite.create_invitation(site, owner, email, :viewer)
+        InviteToSite.invite(site, owner, email, :viewer)
       end
 
       assert {:ok, _} = invite.(site, "i1@example.com")
@@ -110,7 +110,7 @@ defmodule Plausible.Teams.Invitations.InviteToSiteTest do
       add_guest(site2, user: u3, role: :viewer)
 
       invite = fn site, email ->
-        InviteToSite.create_invitation(site, u1, email, :viewer)
+        InviteToSite.invite(site, u1, email, :viewer)
       end
 
       assert {:error, {:over_limit, 3}} = invite.(site, "another@example.com")
@@ -123,7 +123,7 @@ defmodule Plausible.Teams.Invitations.InviteToSiteTest do
       site = new_site(owner: inviter)
 
       assert {:ok, %Plausible.Teams.SiteTransfer{}} =
-               InviteToSite.create_invitation(site, inviter, "vini@plausible.test", :owner)
+               InviteToSite.invite(site, inviter, "vini@plausible.test", :owner)
 
       assert_email_delivered_with(
         to: [nil: "vini@plausible.test"],
@@ -137,7 +137,7 @@ defmodule Plausible.Teams.Invitations.InviteToSiteTest do
       add_member(site.team, user: inviter, role: :admin)
 
       assert {:ok, %Plausible.Teams.SiteTransfer{}} =
-               InviteToSite.create_invitation(site, inviter, "vini@plausible.test", :owner)
+               InviteToSite.invite(site, inviter, "vini@plausible.test", :owner)
 
       assert_email_delivered_with(
         to: [nil: "vini@plausible.test"],
@@ -152,7 +152,7 @@ defmodule Plausible.Teams.Invitations.InviteToSiteTest do
       add_guest(site, user: inviter, role: :editor)
 
       assert {:error, :permission_denied} =
-               InviteToSite.create_invitation(site, inviter, "vini@plausible.test", :owner)
+               InviteToSite.invite(site, inviter, "vini@plausible.test", :owner)
     end
 
     test "allows ownership transfer to existing site members" do
@@ -162,7 +162,7 @@ defmodule Plausible.Teams.Invitations.InviteToSiteTest do
       add_guest(site, user: invitee, role: :viewer)
 
       assert {:ok, %Plausible.Teams.SiteTransfer{}} =
-               InviteToSite.create_invitation(site, inviter, invitee.email, :owner)
+               InviteToSite.invite(site, inviter, invitee.email, :owner)
     end
 
     test "allows creating an ownership transfer even when at team member limit" do
@@ -171,7 +171,7 @@ defmodule Plausible.Teams.Invitations.InviteToSiteTest do
       for _ <- 1..3, do: add_guest(site, role: :viewer)
 
       assert {:ok, _invitation} =
-               InviteToSite.create_invitation(
+               InviteToSite.invite(
                  site,
                  inviter,
                  "newowner@plausible.test",
@@ -186,7 +186,7 @@ defmodule Plausible.Teams.Invitations.InviteToSiteTest do
       add_member(site.team, user: inviter, role: :viewer)
 
       assert {:error, :permission_denied} =
-               InviteToSite.create_invitation(site, inviter, "vini@plausible.test", :viewer)
+               InviteToSite.invite(site, inviter, "vini@plausible.test", :viewer)
     end
 
     test "allows admins to invite editors" do
@@ -195,11 +195,11 @@ defmodule Plausible.Teams.Invitations.InviteToSiteTest do
       add_member(site.team, user: inviter, role: :admin)
 
       assert {:ok, %Plausible.Teams.GuestInvitation{}} =
-               InviteToSite.create_invitation(site, inviter, "vini@plausible.test", :editor)
+               InviteToSite.invite(site, inviter, "vini@plausible.test", :editor)
     end
   end
 
-  describe "bulk_create_invitation/5" do
+  describe "bulk_invite/5" do
     test "initiates ownership transfer for multiple sites in one action" do
       admin_user = new_user()
       new_owner = new_user()
@@ -208,7 +208,7 @@ defmodule Plausible.Teams.Invitations.InviteToSiteTest do
       site2 = new_site(owner: admin_user)
 
       assert {:ok, _} =
-               InviteToSite.bulk_create_invitation(
+               InviteToSite.bulk_invite(
                  [site1, site2],
                  admin_user,
                  new_owner.email,
@@ -238,7 +238,7 @@ defmodule Plausible.Teams.Invitations.InviteToSiteTest do
       site2 = new_site()
 
       assert {:ok, _} =
-               InviteToSite.bulk_create_invitation(
+               InviteToSite.bulk_invite(
                  [site1, site2],
                  superadmin_user,
                  new_owner.email,
