+++ test/plausible/teams/invitations/remove_from_site_test.exs
@@ -13,7 +13,7 @@ defmodule Plausible.Teams.Invitations.RemoveFromSiteTest do
       invite_guest(site, invitee, inviter: inviter, role: :editor)
 
     assert {:ok, removed_invitation} =
-             RemoveFromSite.remove_invitation(invitation.invitation_id, site)
+             RemoveFromSite.remove(invitation.invitation_id, site)
 
     assert removed_invitation.id == invitation.id
     refute Repo.reload(removed_invitation)
@@ -23,7 +23,7 @@ defmodule Plausible.Teams.Invitations.RemoveFromSiteTest do
     site = new_site()
 
     assert {:error, :invitation_not_found} =
-             RemoveFromSite.remove_invitation("does_not_exist", site)
+             RemoveFromSite.remove("does_not_exist", site)
   end
 
   test "does not allow removing invitation from another site" do
@@ -34,7 +34,7 @@ defmodule Plausible.Teams.Invitations.RemoveFromSiteTest do
     invitation = invite_guest(site, invitee, role: :editor, inviter: inviter)
 
     assert {:error, :invitation_not_found} =
-             RemoveFromSite.remove_invitation(invitation.invitation_id, other_site)
+             RemoveFromSite.remove(invitation.invitation_id, other_site)
 
     assert Repo.reload(invitation)
   end
